Formatting and fixes.
