+++ yarn-project/simulator/src/avm/opcodes/external_calls.test.ts
@@ -5,13 +5,12 @@ import { mock } from 'jest-mock-extended';
 
 import { CommitmentsDB, PublicContractsDB, PublicStateDB } from '../../index.js';
 import { AvmContext } from '../avm_context.js';
-import { Field } from '../avm_memory_types.js';
+import { Field, Uint8 } from '../avm_memory_types.js';
 import { initContext } from '../fixtures/index.js';
 import { HostStorage } from '../journal/host_storage.js';
 import { AvmPersistableStateManager } from '../journal/journal.js';
 import { encodeToBytecode } from '../serialization/bytecode_serialization.js';
-import { Return } from './control_flow.js';
-import { Call, StaticCall } from './external_calls.js';
+import { Call, Return, Revert, StaticCall } from './external_calls.js';
 import { Instruction } from './instruction.js';
 import { CalldataCopy } from './memory.js';
 import { SStore } from './storage.js';
@@ -94,7 +93,7 @@ describe('External Calls', () => {
       await instruction.execute(context);
 
       const successValue = context.machineState.memory.get(successOffset);
-      expect(successValue).toEqual(new Field(1n));
+      expect(successValue).toEqual(new Uint8(1n));
 
       const retValue = context.machineState.memory.getSlice(retOffset, retSize);
       expect(retValue).toEqual([new Field(1n), new Field(2n)]);
@@ -182,7 +181,71 @@ describe('External Calls', () => {
 
       // No revert has occurred, but the nested execution has failed
       const successValue = context.machineState.memory.get(successOffset);
-      expect(successValue).toEqual(new Field(0n));
+      expect(successValue).toEqual(new Uint8(0n));
+    });
+  });
+
+  describe('RETURN', () => {
+    it('Should (de)serialize correctly', () => {
+      const buf = Buffer.from([
+        Return.opcode, // opcode
+        0x01, // indirect
+        ...Buffer.from('12345678', 'hex'), // returnOffset
+        ...Buffer.from('a2345678', 'hex'), // copySize
+      ]);
+      const inst = new Return(/*indirect=*/ 0x01, /*returnOffset=*/ 0x12345678, /*copySize=*/ 0xa2345678);
+
+      expect(Return.deserialize(buf)).toEqual(inst);
+      expect(inst.serialize()).toEqual(buf);
+    });
+
+    it('Should return data from the return opcode', async () => {
+      const returnData = [new Fr(1n), new Fr(2n), new Fr(3n)];
+
+      context.machineState.memory.set(0, new Field(1n));
+      context.machineState.memory.set(1, new Field(2n));
+      context.machineState.memory.set(2, new Field(3n));
+
+      const instruction = new Return(/*indirect=*/ 0, /*returnOffset=*/ 0, returnData.length);
+      await instruction.execute(context);
+
+      expect(context.machineState.halted).toBe(true);
+      expect(context.machineState.getResults()).toEqual({
+        reverted: false,
+        output: returnData,
+      });
+    });
+  });
+
+  describe('REVERT', () => {
+    it('Should (de)serialize correctly', () => {
+      const buf = Buffer.from([
+        Revert.opcode, // opcode
+        0x01, // indirect
+        ...Buffer.from('12345678', 'hex'), // returnOffset
+        ...Buffer.from('a2345678', 'hex'), // retSize
+      ]);
+      const inst = new Revert(/*indirect=*/ 0x01, /*returnOffset=*/ 0x12345678, /*retSize=*/ 0xa2345678);
+
+      expect(Revert.deserialize(buf)).toEqual(inst);
+      expect(inst.serialize()).toEqual(buf);
+    });
+
+    it('Should return data and revert from the revert opcode', async () => {
+      const returnData = [new Fr(1n), new Fr(2n), new Fr(3n)];
+
+      context.machineState.memory.set(0, new Field(1n));
+      context.machineState.memory.set(1, new Field(2n));
+      context.machineState.memory.set(2, new Field(3n));
+
+      const instruction = new Revert(/*indirect=*/ 0, /*returnOffset=*/ 0, returnData.length);
+      await instruction.execute(context);
+
+      expect(context.machineState.halted).toBe(true);
+      expect(context.machineState.getResults()).toEqual({
+        reverted: true,
+        output: returnData,
+      });
     });
   });
 });
