+++ yarn-project/simulator/src/avm/serialization/instruction_serialization.ts
@@ -66,7 +66,6 @@ export enum Opcode {
   REVERT,
   KECCAK,
   POSEIDON,
-  // Add new opcodes before this
   SHA256, // temp - may be removed, but alot of contracts rely on it
   PEDERSEN, // temp - may be removed, but alot of contracts rely on it
   TOTAL_OPCODES_NUMBER,
